+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/OldTeleOp.java
@@ -1,17 +1,15 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.HardwareDevice;
 import com.qualcomm.robotcore.hardware.VoltageSensor;
 
 
 @TeleOp
-public class NewTeleOp extends OpMode {
+public class OldTeleOp extends OpMode {
     private DcMotorEx flywheel;
     private DcMotor feedRoller;
     private DcMotor leftDrive;
@@ -19,8 +17,10 @@ public class NewTeleOp extends OpMode {
     private DcMotor rightDrive;
 
     private float flyWheelVelocity = 1300;
-    private float  ticksPerRev = 288;
-    private float offset = 0;
+    private static final int bankVelocity = 1300;
+    private static final int farVelocity = 1900;
+    private static final int maxVelocity = 2200;
+
     private boolean flyWheelPowered;
     private boolean agitatorPowered;
     private boolean feedRollerPowered;
@@ -32,11 +32,8 @@ public class NewTeleOp extends OpMode {
         agitator = hardwareMap.get(CRServo.class, "servo");
         rightDrive = hardwareMap.get(DcMotor.class, "rightDrive");
 
+        // Establishing the direction and mode for the motors
         flywheel.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
-        feedRoller.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        feedRoller.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-
         flywheel.setDirection(DcMotor.Direction.REVERSE);
         feedRoller.setDirection(DcMotor.Direction.REVERSE);
         leftDrive.setDirection(DcMotor.Direction.REVERSE);
@@ -44,7 +41,10 @@ public class NewTeleOp extends OpMode {
         telemetry.addLine("a to turn on/off the flywheel");
         telemetry.addLine("b to turn on/off the agitator");
         telemetry.addLine("x to turn on/off the feed roller");
-        telemetry.addLine("y to turn off flywheel agitator and set feed roller angle");
+        telemetry.addLine("Left bumper for slow speed");
+        telemetry.addLine("right bumper for medium speed");
+
+ 
         telemetry.update();
     }
 
@@ -52,9 +52,6 @@ public class NewTeleOp extends OpMode {
         basicMovement();
         turnOnMotors();
         flyWheel();
-
-        telemetry.addLine("Encoder Position: " + String.valueOf(feedRoller.getCurrentPosition()));
-        telemetry.addLine("Offset: " + offset);
         telemetry.update();
     }
     public double getLowestVoltage() {
@@ -93,29 +90,19 @@ public class NewTeleOp extends OpMode {
             }
         }
         if(gamepad1.xWasPressed()) {
-            if(!feedRoller.isBusy()) {
-                if(feedRollerPowered) {
-                    feedRoller.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-                    feedRollerPowered = false;
-                    feedRoller.setPower(0);
-                } else {
-                    feedRoller.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-                    feedRollerPowered = true;
-                    feedRoller.setPower(1);
-                }
+            if(feedRollerPowered) {
+                feedRollerPowered = false;
+                feedRoller.setPower(0);
+            } else {
+                feedRollerPowered = true;
+                feedRoller.setPower(1);
             }
         }
-        if(gamepad1.yWasPressed()) {
-            float angleOff = (feedRoller.getCurrentPosition() % ticksPerRev);
-
-            feedRollerPowered = false;
-
-            feedRoller.setTargetPosition((int)(feedRoller.getCurrentPosition() - angleOff));
-            feedRoller.setPower(1);
-            feedRoller.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-
-            flywheel.setPower(0);
-            agitator.setPower(0);
+        if(gamepad1.y) {
+            feedRoller.setPower(-0.5);
+        }
+        if(gamepad1.yWasReleased()) {
+            feedRoller.setPower(0);
         }
     }
     public void flyWheel() {
